+++ onnxProtoUtils.hpp
@@ -67,7 +67,7 @@ void deserializeOnnxModel(void const* serializedModel, size_t serializedModelSiz
     // Note: This WARs the very low default size limit (64MB)
     codedInput.SetTotalBytesLimit(std::numeric_limits<int>::max(), std::numeric_limits<int>::max() / 4);
 #endif
-    ONNXTRT_CHECK(model->ParseFromCodedStream(&codedInput) && "Failed to parse the ONNX model.",
+    ONNXTRT_CHECK(model->ParseFromCodedStream(&codedInput), "Failed to parse the ONNX model.",
         ErrorCode::kMODEL_DESERIALIZE_FAILED);
 }